pyproject.toml
[project]
name = "algodz"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "pandas>=2.2.3",
    "numpy>=2.2.5",
    "pybit>=5.11.0",
    "requests>=2.32.3",
    "dotenv>=0.9.9",
    "websocket-client>=1.8.0",
    "sortedcontainers>=2.4.0",
    "aiohttp>=3.12.0",
    "ujson>=5.10.0",
    "python-dotenv>=1.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/algodz"]
